feat(web): group dashboard settings; split PWA & JSON backup
- New DashboardSection combines hints + density ("Вигляд") and
  active modules + reorder + hide-inactive ("Модулі дашборду") into
  one collapsible group
- GeneralSection slimmed to Онбординг + Хмарна синхронізація
- New PWASection and DataExportSection registered under Додатково
- Rename «Резервна копія Hub» → «Експорт/імпорт JSON»; clarify that
  cloud sync is primary, JSON is for users w/o account or portability
- Replace inline KVStore wrapper / raw localStorage in settings with
  shared webKVStore + safeReadStringLS/safeWriteLS helpers
